Add files via upload
This project focuses on the factors that will impact the baby weight when were born. From the original data provided, a total of seventeen independent vari- ables are collected for further analysis on the impact on birthweight. They are parents’ age, parents’ education level, parents’ races, month prenatal care began, the total number of prenatal visits, one-minute apgar score, five-minute apgar score, average drink, and cigarettes consumption, and gender of the baby. We assumed that the higher the birth weight is, the healthier the babies are. This report will examine the impact of each variable on birthweight and the change of impact after regrouping and classification of certain variables.